requirements.txt
youtube_dl
requests
moviepy
pillow
pyttsx3
pydrive
pypiwin32